% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoder.R
\name{print.fhir_tx_encoder}
\alias{print.fhir_tx_encoder}
\title{Summarise a FHIR terminology encoder}
\usage{
\method{print}{fhir_tx_encoder}(x, ...)
}
\arguments{
\item{x}{An object of class `fhir_tx_encoder`.}

\item{...}{Unused, present for compatibility with the `print()` generic.}
}
\value{
`x`, invisibly.
}
\description{
Summarise a FHIR terminology encoder
}
\examples{
\dontrun{
encoder <- fhir_tx_encoder(
  scope = "http://snomed.info/sct?fhir_vs=ecl/(\%3E\%3E\%20363346000)"
)
print(encoder)
}
}
